Add possibility of selecting frame to process
Add the possibility to select the spin configuration to process
with vdc utility via "vdc intial final". If no option is given,
the code will run as in the previous version. Also, if only the
"initial" frame is given, the code will run from "initial" to
the last in the directory.
In addition added Eclispe IDE project files and folder extensions
in .gitignore files