Modify hamr simlation program
Modify hamr simulation program by creating two simulation modes:
pulsed and continuous. In pulsed mode the write head and laser
are fixed at the centre of the bit and the laser switched on
following a Gaussian time profile.  In continuous mode the
write head sweeps across the track and the laser is on
throughout the whole simulation process.
The thermal profile is Gaussian in space with possibility of
specifying different FWHM in x,y-directions for both modes.
The applied field is trapezoidal in time and uniform in the space
region where is applied for both modes.
New input parameters have been defined to allow the control of the
hamr simulations.